ItemSearch 1.7
- [Fitting Room]  Try on another player's whole outfit from the examine window.
- [Fitting Room] Rework internal structures to hopefully be less unstable.
- [Fitting Room] Show previews of saved outfits
- [Fitting Room] Avoid spamming chat when switching outfits
- [Fitting Room] Add command to open fitting room (`/fittingroom`)
- [Fitting Room] Clear old outfit when switching using saved outfits (Hold shift to use old "Merge" style)
- [Fitting Room] Undelete. Hold Shift
- [Search] Togglable display of advanced filters.
- [Search] Tag based filtering

Tag Based Filtering:
Using the "Search" input you can now search apply other filters by using "Tags". These are simple strings such as `[not dyeable]` or `[ilvl:50]` that can be combined in the search input to apply filters. When using a tag, the filter that it is applied to will be highlghted in blue and will not be changable in the normal way until the tag is cleared.
https://github.com/Caraxi/ItemSearchPlugin/blob/master/tag-filtering.md