TextToTalk v1.4
This version adds the VoiceUnlocker function to the plugin configuration, which allows you to enable system voices that don't show up in the list by default. While this won't enable *every* voice you may have installed, it will enable the voices designated as mobile-only (which is most of them). Using this function modifies your registry, so do read the big warning that pops up when you click the new button.