Update DiscordBridge
This update now includes the ability to register a DM with the bot as a chat channel.

Registering a chat channel works the same as normal discord server channels, via `xl!setchannel`

Relayed messages have been converted to embeds because webhooks aren't allowed in DMs. You may remember this layout from the anycient build-in discord relay.