[TESTING] XIVDeck 0.0.1 (by KazWolfe)
XIVDeck is a plugin that allows for first-class integration between the
Elgato Stream Deck and Final Fantasy XIV. This plugin is divided into
two component parts: the game plugin (as included here) and the Stream
Deck Plugin (available on the plugin's GitHub repository).

This plugin is currently pre-release and requires testing for bugs and
general issues that may affect usability and maintainability in the long
run. While I have attempted to add some degree of safety checking to
ensure that the game doesn't send illegal data, the nature of this
plugin does present a possibility that a user can allow such illegal
data to be generated and sent.

Plugin GitHub Repository: https://github.com/KazWolfe/XIVDeck