Tidy Chat v0.1.1.1
**Additions**

- Adds setting to display Instance # in the DTR Bar. It is a separate option from `Better /instance message` and you can even hide the instance message altogether under `System->Not "Spam"->[] Hide /instance message`

Upon first enabling the DTR Bar you will need to use `/instance` if an instanced area or switch to an instanced area to set the current instance. If you leave the setting enabled it will always display the instance without the need to use `/instance`. This is because it detects the instance via the chat message that appears when you enter an instanced area and not through any game data.

**Improvements**

- Better detection of instance number by capturing the actual 1/2/3 text blocks with a regex capture group instead of jankily parsing strings. Honestly this should have been done ages ago but I forgot to update it