DiscordBridge bugfix
Fixed type number for `fcannounce` and `fclogin`. 

Added `fcrank` too.

Removing the `sign` type because I haven't validated it and it probably doesn't work. Preferably, I'd like to switch this to some shared package of chat type definitions so these don't need to be done by hand for each specific type. (And because FFXIV uses a number of grouped types, which will not be supported at this t me).

You will need to use `xl!unsetchannel` and `xl!setchannel` again to update these types.