DailyDuty v1.5.0.3
*No Code Changes*

This update is to inform you that the next version of DailyDuty will be a complete-rebuild, as such all your character-specific settings will be reset.

If you wish to keep a copy of your current settings please backup `%appdata%\XIVLauncher\pluginConfigs\DailyDuty.json`

More generalized configurations such as Timers Windows, and Todo Windows will not be effected.