[Testing] FFXIV2MQTT 1.0.1.0
Can now control if you want topics that include timers (e.g. WHM Job Gauge Lily Timer) to be periodically sent.
Updated the config screen to have a little bit more help in it.
Fixed the logged in event not having character name.
Added ID's to a couple topics so you can match against the ID instead of the name if you would like. More to come in the future.